release v 1.0 create def user client with protection

+++ src/main/java/com/example/amrachat_usr_clnt_api/controller/ClientsController.java
@@ -54,6 +54,10 @@ public class ClientsController {
     public ResponseEntity<?> deleteUser(@RequestBody Client client) {
         String resMsg;
         try {
+            if(client.getId()==1){
+                throw new Exception("Permission denied for delete client with id = 1");
+            }
+
             if(clientsDao.existsById(client.getId())) {
                 clientsDao.delete(client);
                 resMsg = "{Client with id: " + client.getId() + " deleted}";
